feat(api): extract TransporterRulesService and refactor controller
- Handles complex transactions for flow nodes/edges and fee rules
- Maps temporary node IDs during save operations
- Reduces controller from ~400 lines to ~50 lines
